requirements.txt
# Aura Lead Hunter - Dependencies
# Python 3.11+

# Telegram UserBot API
telethon==1.36.0

# AI/LLM Integration
openai==1.58.1
httpx==0.28.1

# Data Processing
pandas==2.2.3

# Environment & Configuration
python-dotenv==1.0.1

# Async utilities
aiofiles==24.1.0

# Rich console output for debugging
rich==13.9.4
